dape map/start: new logics for pangenomes
For distant genomes (i.e. genomes from distant species) there are many
examples of convergent evolution on some reactions; reactions flagged as dispensable
were indeed conserved, meaning that the genes are not orthologs but their function
is similar in the KEGG map.
The new logic will consider the presence of the reaction IDs instead of being ortholog-centric.
Also the phenotypic variability will consider the reaction IDs.
For strain belonging to the same species the results may be unaltered.